refactor: Address PR review comments
